fix(acl): 权限位按无符号 32 位语义传递
ServerSync 的 uint64 只取低 32 位，uint32 字段拓宽为 Long，避免 toInt() 把 bit31 变成负数。

+++ app/src/main/java/dev/woms/mumdroid/core/model/UserModeration.kt
@@ -144,7 +144,7 @@ object UserModeration {
      * `Move | Write`.
      */
     fun canResetUserContent(
-        rootPermissions: Int,
+        rootPermissions: Long,
         versionV2: Long,
         legacyVersion: Int = 0,
     ): Boolean {
